src/backend/pyproject.toml
[tool.poetry]
name = "todo-app-backend"
version = "0.1.0"
description = ""
authors = ["Joaquín Neira <joaquin.neiraj@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
aiopath = "^0.6.11"
dependency-injector = "^4.41.0"
flask = {extras = ["async"], version = "^2.3.3"}
webargs = "^8.2.0"
flask-cors = "^3.0.10"
sqlalchemy = "^2.0.4"
psycopg = { extras = ["binary"], version = "^3.1.8" }
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}


[tool.poetry.group.test.dependencies]
pytest = "^7.2.1"
pytest-factoryboy = "^2.5.1"
freezegun = "^1.2.2"
pytest-mock = "^3.10.0"


[tool.poetry.group.development.dependencies]
black = "^23.7.0"
ruff = "^0.0.284"
pyright = "^1.1.322"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Formatting
[tool.black]
line-length = 79
skip-magic-trailing-comma = true

# Testing
[tool.pytest.ini_options]
testpaths = ["tests"]

# typing
[tool.pyright]
pythonversion = "3.11"

[tool.ruff]
ignore = [
    "ANN003", #Missing type annotation for `**kwargs` 
    "ANN101", #Missing type annotation for `self` in method
    "ANN102", #Missing type annotation for `cls` in classmethod
    "ANN204", #Missing return type annotation for special method `__init__`
    "BLE001", #Do not catch blind exception: `Exception`
    "COM812", #Trailing comma missing
    "D100",   #Missing docstring in public module
    "D101",   #Missing docstring in public class
    "D102",   #Missing docstring in public method
    "D103",   #Missing docstring in public function
    "D104",   #Missing docstring in public package
    "D105",   #Missing docstring in magic method
    "D106",   #Missing docstring in public nested class
    "D107",   #Missing docstring in `__init__`
    "D203",   #1 blank line required before class docstring
    "D204",   #1 blank line required after class docstring
    "D212",   #Multi-line docstring summary should start at the first line
    "RET503", #Missing explicit `return` at the end of function able to return non-`None` value
]
select = ["ALL"]
line-length = 79
show-fixes = true
show-source = true
target-version = "py311"

[tool.ruff.flake8-import-conventions]
[tool.ruff.flake8-import-conventions.extend-aliases]
"datetime" = "dt"

[tool.ruff.flake8-pytest-style]
fixture-parentheses = false

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.isort]
combine-as-imports = true
known-first-party = ["app", "tests"]

[tool.ruff.mccabe]
max-complexity = 5
